move utility functions for deduction script into own file
