+++ databases/catdat/scripts/deduce-category-properties.ts
@@ -1,24 +1,14 @@
 import { SqliteError, type Database } from 'better-sqlite3'
+import { get_assumption_string, get_conclusion_string, is_subset } from './shared'
 import {
-	get_assumption_string,
-	get_conclusion_string,
-	is_subset,
+	CategoryMeta,
+	CategoryPropertyMeta,
+	get_all_decided_properties,
+	get_categories,
+	get_normalized_category_implications,
+	get_properties_dict,
 	NormalizedCategoryImplication,
-} from './shared'
-
-type CategoryMeta = {
-	id: string
-	name: string
-	dual_category_id: string | null
-}
-
-type CategoryPropertyMeta = {
-	id: string
-	dual_property_id: string | null
-	relation: string
-	negation: string
-	conditional: string
-}
+} from './categories.utils'
 
 /**
  * Deduce properties of categories from given ones
@@ -96,133 +86,11 @@ export function deduce_category_properties(db: Database) {
 }
 
 /**
- * Implications have the form:
- *
- * P_1 + ... + P_n ----> Q_1 + ... + Q_m
- *
- * or
- *
- * P_1 + ... + P_n <---> Q_1 + ... + Q_m
- *
- * This function decomposes them into normalized implications,
- * which have the form:
- *
- * P_1 + ... + P_n ----> Q
- */
-function get_normalized_category_implications(
-	db: Database,
-): NormalizedCategoryImplication[] {
-	const all_implications_db = db
-		.prepare(
-			`SELECT
-				v.id,
-				v.assumptions,
-				v.conclusions,
-				v.is_equivalence
-			FROM category_implications_view v`,
-		)
-		.all() as {
-		id: string
-		assumptions: string
-		conclusions: string
-		is_equivalence: number
-	}[]
-
-	const implications: NormalizedCategoryImplication[] = []
-
-	for (const impl of all_implications_db) {
-		const assumptions: string[] = JSON.parse(impl.assumptions)
-		const conclusions: string[] = JSON.parse(impl.conclusions)
-
-		for (const conclusion of conclusions) {
-			implications.push({
-				id: impl.id,
-				assumptions: new Set(assumptions),
-				conclusion,
-			})
-		}
-
-		if (impl.is_equivalence) {
-			for (const assumption of assumptions) {
-				implications.push({
-					id: impl.id,
-					assumptions: new Set(conclusions),
-					conclusion: assumption,
-				})
-			}
-		}
-	}
-
-	return implications
-}
-
-/**
- * Returns the list of categories saved in the database.
- */
-function get_categories(db: Database) {
-	return db
-		.prepare(
-			`SELECT id, name, dual_category_id
-			FROM categories ORDER BY lower(name)`,
-		)
-		.all() as CategoryMeta[]
-}
-
-/**
- * Returns a dictionary of properties saved in the database.
- */
-function get_properties_dict(db: Database) {
-	const properties = db
-		.prepare(
-			`SELECT
-				p.id, p.dual_property_id, p.relation,
-				r.negation, r.conditional
-			FROM category_properties p
-			INNER JOIN relations r ON r.relation = p.relation
-			ORDER BY lower(p.id)`,
-		)
-		.all() as CategoryPropertyMeta[]
-
-	const dict: Record<string, CategoryPropertyMeta> = {}
-
-	for (const p of properties) dict[p.id] = p
-
-	return dict
-}
-
-/**
  * Clears all the deduced properties.
  * This runs before the deduction starts.
  */
 function delete_deduced_category_properties(db: Database) {
 	db.prepare(`DELETE FROM category_property_assignments WHERE is_deduced = TRUE`).run()
-}
-
-/**
- * Returns a dictionary with all properties that are satisfied or unsatisfied,
- * grouped by category and value.
- */
-function get_all_decided_properties(db: Database, categories: { id: string }[]) {
-	const rows = db
-		.prepare(
-			`SELECT property_id, category_id, is_satisfied
-			FROM category_property_assignments`,
-		)
-		.all() as { property_id: string; category_id: string; is_satisfied: boolean }[]
-
-	const grouped: Record<string, { satisfied: Set<string>; unsatisfied: Set<string> }> =
-		{}
-
-	for (const category of categories) {
-		grouped[category.id] = { satisfied: new Set(), unsatisfied: new Set() }
-	}
-
-	for (const row of rows) {
-		const { property_id, category_id, is_satisfied } = row
-		grouped[category_id][is_satisfied ? 'satisfied' : 'unsatisfied'].add(property_id)
-	}
-
-	return grouped
 }
 
 /**
